refactor: :building_construction: move buildServiceContexts() to a more sensible file
